Remove redundant styling
`/source/stylesheets/screen.css.scss` had a bunch of hard-coded styling
which was breaking the build because of undefined colour names.

Removing the extra styling fixes the build locally.

That styling came as part of the initial commit so it's not clear what
purpose it serves now.

The gem uses the Design System and should not be using all this extra
styling.