Remove unused deploy script
This repo is deployed using the CloudFoundry Travis provider, as defined in `.travis.yml`.

This deploy script was erroneously included in the initial commit (1bebd05) as this repo appears to have been created from a copy of the registers tech docs, where deployments are done using a script provider in order to provide a zero-downtime deploy.