Update table markdown guidance
This removes the requirement to add a `<div>` above and below tables. This was originally required because omitting them would break the table completely, but this no longer seems to be an issue. The spacing between the table and the surrounding content is now slightly reduced as a result, but this should be fixed with CSS in the tech docs gem, rather than with non-semantic content.

This commit also documents new accessibility guidance whereby tables with 3 or more columns need a row header, introduced in https://github.com/alphagov/tech-docs-gem/pull/214.

While I'm here, I've updated the markdown example to use padded cells for horizontal alignment, which is good practice if there is room to do so. I've also omitted the `|:---|` left-alignment instruction, since cells are left-aligned by default. Most markdown table examples you'll come across in the wild use the default `|---|` separator.