Use Node.js version 14.x.x for Travis CI
Travis CI is currently defaulting to using Node.js v8.x.x because the
Travis config file does not specify to use Node at all.

However a version of Node greater than 10 is needed for this repo,
because the tech-docs-gem uses it.

This commit uses nvm to make sure that the version of Node used for
tests is in the current long term support (LTS) series.