Update docs to prefer 'main' over 'master'
We're trying to phase out use of the word 'master' as the default
default branch name, along with the rest of the software engineering
community [[1]]. This commit updates the docs, as we're updating the
template in PR alphagov/tech-docs-template#228. Also add a note about
changing the default in a future release of the gem.

[1]: github.com/renaming